+++ lib/src/connection_limiter.rs
@@ -131,6 +131,7 @@ mod tests {
             password: password.into(),
             max_http2_conns: None,
             max_http3_conns: None,
+            max_traffic_bytes: None,
         }
     }
 
@@ -145,6 +146,7 @@ mod tests {
             password: password.into(),
             max_http2_conns: h2,
             max_http3_conns: h3,
+            max_traffic_bytes: None,
         }
     }