+++ lib/tests/authentication.rs
@@ -135,6 +135,7 @@ async fn run_endpoint(
                 password: "b".into(),
                 max_http2_conns: None,
                 max_http3_conns: None,
+                max_traffic_bytes: None,
             },
         )));
     }
@@ -369,6 +370,7 @@ async fn run_endpoint_with_conn_limit(listen_address: &SocketAddr, max_http2_con
             password: "b".into(),
             max_http2_conns: Some(max_http2_conns),
             max_http3_conns: None,
+            max_traffic_bytes: None,
         }])
         .build()
         .unwrap();